requirements.txt
pyyaml
pydantic
ipython
rich